refactor: move feature strategies store type
