+++ src/lib/features/playground/offline-unleash-client.ts
@@ -1,6 +1,6 @@
 import { SdkContextSchema } from 'lib/openapi/spec/sdk-context-schema';
 import { InMemStorageProvider, FeatureEvaluator } from './feature-evaluator';
-import { FeatureConfigurationClient } from 'lib/types/stores/feature-strategies-store';
+import { FeatureConfigurationClient } from 'lib/features/feature-toggle/types/feature-strategies-store-type';
 import { Segment } from './feature-evaluator/strategy/strategy';
 import { ISegment } from 'lib/types/model';
 import { serializeDates } from '../../types/serialize-dates';
